+++ src/sql/connection/connection.py
@@ -26,7 +26,6 @@ from ploomber_core.exceptions import modify_exceptions
 
 
 from sql.store import store
-from sql.telemetry import telemetry
 from sql import exceptions, display
 from sql.error_handler import handle_exception
 from sql.parse import (
@@ -990,13 +989,7 @@ class DBAPIConnection(AbstractConnection):
 
     is_dbapi_connection = True
 
-    @telemetry.log_call("DBAPIConnection", payload=True)
-    def __init__(self, payload, connection, alias=None, config=None):
-        try:
-            payload["engine"] = type(connection)
-        except Exception as e:
-            payload["engine_parsing_error"] = str(e)
-
+    def __init__(self, connection, alias=None, config=None):
         # detect if the engine is a native duckdb connection
         _is_duckdb_native = _check_if_duckdb_dbapi_connection(connection)
 
@@ -1084,12 +1077,7 @@ class DBAPIConnection(AbstractConnection):
 class SparkConnectConnection(AbstractConnection):
     is_dbapi_connection = False
 
-    @telemetry.log_call("SparkConnectConnection", payload=True)
-    def __init__(self, payload, connection, alias=None, config=None):
-        try:
-            payload["engine"] = type(connection)
-        except Exception as e:
-            payload["engine_parsing_error"] = str(e)
+    def __init__(self, connection, alias=None, config=None):
         self._driver = None
 
         # TODO: implement the dialect blacklist and add unit tests