+++ src/sql/plot.py
@@ -29,7 +29,6 @@ except ModuleNotFoundError:
     np = None
 
 import sql.connection
-from sql.telemetry import telemetry
 import warnings
 
 
@@ -178,10 +177,7 @@ def _boxplot_stats(conn, table, column, whis=1.5, autorange=False, with_=None):
 
 # https://github.com/matplotlib/matplotlib/blob/ddc260ce5a53958839c244c0ef0565160aeec174/lib/matplotlib/axes/_axes.py#L3915
 @requires(["matplotlib"])
-@telemetry.log_call("boxplot", payload=True)
-def boxplot(
-    payload, table, column, *, orient="v", with_=None, conn=None, ax=None, schema=None
-):
+def boxplot(table, column, *, orient="v", with_=None, conn=None, ax=None, schema=None):
     """Plot boxplot
 
     Parameters
@@ -230,8 +226,6 @@ def boxplot(
     """
     if not conn:
         conn = sql.connection.ConnectionManager.current
-
-    payload["connection_info"] = conn._get_database_information()
 
     _table = enclose_table_with_double_quotations(table, conn)
     if schema:
@@ -316,9 +310,7 @@ def _get_bar_width(ax, bins, bin_size, binwidth):
 
 
 @requires(["matplotlib"])
-@telemetry.log_call("histogram", payload=True)
 def histogram(
-    payload,
     table,
     column,
     bins,
@@ -409,7 +401,7 @@ def histogram(
         _table = f'"{schema}"."{_table}"'
 
     ax = ax or plt.gca()
-    payload["connection_info"] = conn._get_database_information()
+
     if category:
         if isinstance(column, list):
             if len(column) > 1:
@@ -874,9 +866,7 @@ def _bar(table, column, with_=None, conn=None):
 
 
 @requires(["matplotlib"])
-@telemetry.log_call("bar", payload=True)
 def bar(
-    payload,
     table,
     column,
     show_num=False,
@@ -928,7 +918,6 @@ def bar(
         _table = f'"{schema}"."{_table}"'
 
     ax = ax or plt.gca()
-    payload["connection_info"] = conn._get_database_information()
 
     if column is None:
         raise exceptions.UsageError("Column name has not been specified")
@@ -1059,9 +1048,7 @@ def _pie(table, column, with_=None, conn=None):
 
 
 @requires(["matplotlib"])
-@telemetry.log_call("bar", payload=True)
 def pie(
-    payload,
     table,
     column,
     show_num=False,
@@ -1107,7 +1094,6 @@ def pie(
         _table = f'"{schema}"."{_table}"'
 
     ax = ax or plt.gca()
-    payload["connection_info"] = conn._get_database_information()
 
     if column is None:
         raise exceptions.UsageError("Column name has not been specified")